# SIMD: Replace SVML/ASM of tanh(f32, f64) with universal intrinsics