@r-devulap 2021-11-13
Nice, will review it over the next few days. The `test_umath_accuracy` tests run only on x86_64 and on CPU's with AVX-512, we might have to find a way to run them on other platforms too.