@seiko2plus 2021-11-13
> At some point this should get a benchmark

I still need to introduce other two universal intrinsics (npyv_all_##sfx, npyv_any_##sfx) to speed up boolean tests across the vector elements to speed up falling back to `C` on `ppc64le` and `arm64` then I'm gonna publish benchmark result and set it ready for review.

>  There are ULP tests in test_umath_accuracy

I just converted the x86/avx512 instructions of SMVL into C intrinsics, I did little tweaks but still almost the same implementation.

> test_umath_accuracy tests run only on x86_64 and on CPU's with AVX-512

I thought `test_umath_accuracy` is enabled with `AVX2 & FMA3` too. since the universal intrinsics passed with one platform it should pass with the others too and also the current code only enabled for SIMD extensions with fused support.