@r-devulap 2021-11-13
> I thought `test_umath_accuracy` is enabled with `AVX2 & FMA3` too. since the universal intrinsics passed with one platform it should pass with the others too and also the current code only enabled for SIMD extensions with fused support.

My bad, it is enabled for `AVX2 & FMA3` too. And yes, fma instruction is a must, otherwise the accuracy can get really bad.