@seiko2plus 2021-11-15
> I still need to introduce other two universal intrinsics (npyv_all_##sfx, npyv_any_##sfx) to speed up boolean tests across

take it back, no performance changes `npyv_tobits` almost perform the same in this case, however, I will need to implement them for other cases in different pr one day.

@mattip,
>  At some point this should get a benchmark

done