@seiko2plus 2021-11-15
github still don't offer an option to change the pr branch, no way to replace `seiko2plus:svml2npyv/tanh_f32` to `seiko2plus:svml2npyv/tanh`, the plan was just to replace single precision but now I replaced both.