@mattip 2021-11-15
I am a bit surprised that the benchmarks improve AVX512 performance. Isn't it using SVML?