@seiko2plus 2021-11-15
@mattip, depending on the compiler, the latest versions of clang and GCC may even perform better, the current benchmark only shows improvements on `avx512` for non-contiguous only, assuming SVML compiled against the same compiler still inlining can reduce extra load/store over the stack. There's already an open discussion on this matter, https://www.mail-archive.com/numpy-discussion@python.org/msg05645.html.