@seiko2plus 2022-01-19
@r-devulap,
> I would suggest getting rid of the C callout for special cases (large numbers, INF and NAN)

done for single-precision.