@r-devulap 2022-01-28
> @r-devulap,
> 
> > I would suggest getting rid of the C callout for special cases (large numbers, INF and NAN)
> 
> done for single-precision.

Once its fixed for double precision, this PR will be good to merge.