@seiko2plus 2022-02-01
@r-devulap, it's now for both single & double. I have revisited the assembly code again and I just realized that there's no special handling when `|x| > HUGE_THRESHOLD` just returning `+-1` even for double-precision. unfortunately, SVML/tanh description wasn't clear enough and asm paths of the callout were strange a little bit.