@rgommers 2022-11-15
I just went on a little goose chase to find out to the `svml_filter` content added in this PR, so let me cross-link it for future reference.
- This PR adds `float32` and `float64` implementations of `tanh`, and therefore filters out the relevant object files from SVML
- gh-21955 adds a `float16 implementation in SVML, and removes the contents of `svml_filter` to include the two object files that are now needed again in the build
  - It does leave the empty filter and the code comment in place, which is now confusing.
  - In the `meson` branch, I will just get rid of that, but instead put a one line note that filenames which are no longer needed because they're fully implemented with universal intrinsics should be commented out.