@seiko2plus 2022-11-16
@rgommers, #21955 didn't reuse the universal intrinsic implementation to handle half-precision over single, instead, it brings SVML objects back. However, @r-devulap should have removed only `svml_z0_tanh_s_la.s` and kept `svml_z0_tanh_d_la.s`.

So you can choose between filtering out `svml_z0_tanh_d_la.s`, or ignoring its existence. In general, the reason behind this filteration is to reduce the binary size.